+++ src/usecases/fabricas/fabrica_usecases.py
@@ -1,7 +1,7 @@
 from src.repositorios.mock.armazenamento_usuario_volatil import ArmazenamentoUsuarioVolatil
 
 from src.usecases.uc_adicionar_email import UCAdicionarEmail
-from src.usecases.uc_get_por_telefone import UCGetPorTelefone
+from src.usecases.uc_get_usuario_por_telefone import UCGetUsuarioPorTelefone
 from src.usecases.uc_remover_email import UCRemoverEmail
 from src.usecases.uc_editar_email import UCEditarEmail
 from src.usecases.uc_adicionar_endereco import UCAdicionarEndereco
@@ -12,8 +12,8 @@ from src.usecases.uc_remover_telefone import UCRemoverTelefone
 from src.usecases.uc_editar_telefone import UCEditarTelefone
 from src.usecases.uc_cadastrar_usuario import UCCadastrarUsuario
 from src.usecases.uc_deletar_usuario_por_email import UCDeletarUsuarioPorEmail
-from src.usecases.uc_get_por_userid import UCGetPorUserId
-from src.usecases.uc_get_por_email import UCGetPorEmail
+from src.usecases.uc_get_usuario_por_userid import UCGetUsuarioPorUserId
+from src.usecases.uc_get_usuario_por_email import UCGetUsuarioPorEmail
 
 
 class FabricaUsecases:
@@ -56,10 +56,10 @@ class FabricaUsecases:
         return UCDeletarUsuarioPorEmail(self.repo)
 
     def getUsuarioPorUserId(self):
-        return UCGetPorUserId(self.repo)
+        return UCGetUsuarioPorUserId(self.repo)
 
     def getUsuarioPorEmail(self):
-        return UCGetPorEmail(self.repo)
+        return UCGetUsuarioPorEmail(self.repo)
 
     def getUsuarioPorTelefone(self):
-        return UCGetPorTelefone(self.repo)
+        return UCGetUsuarioPorTelefone(self.repo)
