+++ src/usecases/uc_get_usuario_por_email.py
@@ -1,19 +1,17 @@
 from src.interfaces.IRepoUsuario import IArmazenamento
 from src.usecases.erros.erros_uc_alteracao_info_cadastro import ErroUsuarioNaoExiste
-from src.usecases.erros.erros_usecase import ErroIdInvalido
 
 
-class UCGetPorUserId:
+class UCGetUsuarioPorEmail:
     usuarioRepo: IArmazenamento
 
     def __init__(self, usuarioRepo: IArmazenamento):
         self.usuarioRepo = usuarioRepo
 
-    def __call__(self, userId: int):
-        if(userId < 0):
-            raise ErroIdInvalido
-        #Erro é re-levantado pois repo pode ser alterado
+    def __call__(self, email: str):
+        # Email nao pode ser validado pelo model de email - nao será feito erro de badrequest
+        # Erro é re-levantado pois repo pode ser alterado
         try:
-            return self.usuarioRepo.getUsuarioPorUserId(userId)
+            return self.usuarioRepo.getUsuarioPorEmail(email)
         except:
             raise ErroUsuarioNaoExiste
