+++ src/usecases/uc_get_usuario_por_telefone.py
@@ -2,16 +2,15 @@ from src.interfaces.IRepoUsuario import IArmazenamento
 from src.usecases.erros.erros_uc_alteracao_info_cadastro import ErroUsuarioNaoExiste
 
 
-class UCGetPorEmail:
+class UCGetUsuarioPorTelefone:
     usuarioRepo: IArmazenamento
 
     def __init__(self, usuarioRepo: IArmazenamento):
         self.usuarioRepo = usuarioRepo
 
-    def __call__(self, email: str):
-        # Email nao pode ser validado pelo model de email - nao será feito erro de badrequest
-        # Erro é re-levantado pois repo pode ser alterado
+    def __call__(self, ddd: int, numero: str):
+        #Erro é re-levantado pois repo pode ser alterado
         try:
-            return self.usuarioRepo.getUsuarioPorEmail(email)
+            return self.usuarioRepo.getUsuarioPorTelefone(ddd, numero)
         except:
             raise ErroUsuarioNaoExiste
