+++ src/usecases/uc_get_usuario_por_userid.py
@@ -1,16 +1,19 @@
 from src.interfaces.IRepoUsuario import IArmazenamento
 from src.usecases.erros.erros_uc_alteracao_info_cadastro import ErroUsuarioNaoExiste
+from src.usecases.erros.erros_usecase import ErroIdInvalido
 
 
-class UCGetPorTelefone:
+class UCGetUsuarioPorUserId:
     usuarioRepo: IArmazenamento
 
     def __init__(self, usuarioRepo: IArmazenamento):
         self.usuarioRepo = usuarioRepo
 
-    def __call__(self, ddd: int, numero: str):
+    def __call__(self, userId: int):
+        if(userId < 0):
+            raise ErroIdInvalido
         #Erro é re-levantado pois repo pode ser alterado
         try:
-            return self.usuarioRepo.getUsuarioPorTelefone(ddd, numero)
+            return self.usuarioRepo.getUsuarioPorUserId(userId)
         except:
             raise ErroUsuarioNaoExiste
